Trait gooey_renderer::Renderer

pub trait Renderer: Debug + Send + Sync + Sized + 'static {

    fn size(&self) -> Size2D<f32, Points>;

    fn theme(&self) -> SystemTheme;

    fn clip_to(&self, bounds: Rect<f32, Points>) -> Self;

    fn clip_bounds(&self) -> Rect<f32, Points>;

    fn scale(&self) -> Scale<f32, Points, Pixels>;

    fn render_text(
        &self, 
        text: &str, 
        baseline_origin: Point2D<f32, Points>, 
        options: &TextOptions
    );

    fn measure_text(
        &self, 
        text: &str, 
        options: &TextOptions
    ) -> TextMetrics<Points>;

    fn fill_rect(&self, rect: &Rect<f32, Points>, color: Color);

    fn stroke_rect(&self, rect: &Rect<f32, Points>, options: &StrokeOptions);

    fn stroke_line(
        &self, 
        point_a: Point2D<f32, Points>, 
        point_b: Point2D<f32, Points>, 
        options: &StrokeOptions
    );

    fn bounds(&self) -> Rect<f32, Points> { ... }

    fn render_text_with_style<F: FallbackComponent<Value = ColorPair>>(
        &self, 
        text: &str, 
        baseline_origin: Point2D<f32, Points>, 
        style: &Style
    ) { ... }

    fn measure_text_with_style(
        &self, 
        text: &str, 
        style: &Style
    ) -> TextMetrics<Points> { ... }

    fn fill_rect_with_style<F: FallbackComponent<Value = ColorPair>>(
        &self, 
        rect: &Rect<f32, Points>, 
        style: &Style
    ) { ... }

    fn stroke_rect_with_style<F: FallbackComponent<Value = ColorPair>>(
        &self, 
        rect: &Rect<f32, Points>, 
        style: &Style
    ) { ... }

    fn stroke_line_with_style<F: FallbackComponent<Value = ColorPair>>(
        &self, 
        point_a: Point2D<f32, Points>, 
        point_b: Point2D<f32, Points>, 
        style: &Style
    ) { ... }

}

Implements drawing APIs.

Required methods

fn size(&self) -> Size2D<f32, Points>

The size of the area being drawn.

fn theme(&self) -> SystemTheme

Returns the current system theme.

fn clip_to(&self, bounds: Rect<f32, Points>) -> Self

Returns a new renderer instance with the state such that each operation executes as if the origin is bounds.origin. The returned instance’s size() should equal bounds.size.

fn clip_bounds(&self) -> Rect<f32, Points>

A [Rect] representing the area being drawn. This rect should be offset relative to the origin of the renderer.

fn scale(&self) -> Scale<f32, Points, Pixels>

The scale when converting between Points and Pixels.

fn render_text(
&self,
text: &str,
baseline_origin: Point2D<f32, Points>,
options: &TextOptions
)

Renders text at baseline_origin with options.

fn measure_text(&self, text: &str, options: &TextOptions) -> TextMetrics<Points>

Measures text using options.

fn fill_rect(&self, rect: &Rect<f32, Points>, color: Color)

Fills rect using color.

fn stroke_rect(&self, rect: &Rect<f32, Points>, options: &StrokeOptions)

Strokes the outline of rect using options.

fn stroke_line(
&self,
point_a: Point2D<f32, Points>,
point_b: Point2D<f32, Points>,
options: &StrokeOptions
)

Draws a line between point_a and point_b using options.

Loading content...
Provided methods

fn bounds(&self) -> Rect<f32, Points>

A [Rect] representing the area being drawn. Due to how rendering works, the origin is always zero.

fn render_text_with_style<F: FallbackComponent<Value = ColorPair>>(
&self,
text: &str,
baseline_origin: Point2D<f32, Points>,
style: &Style
)

Renders text at baseline_origin with options.

fn measure_text_with_style(
&self,
text: &str,
style: &Style
) -> TextMetrics<Points>

Measures text using style.

fn fill_rect_with_style<F: FallbackComponent<Value = ColorPair>>(
&self,
rect: &Rect<f32, Points>,
style: &Style
)

Fills rect using style.

fn stroke_rect_with_style<F: FallbackComponent<Value = ColorPair>>(
&self,
rect: &Rect<f32, Points>,
style: &Style
)

Strokes the outline of rect using style.

fn stroke_line_with_style<F: FallbackComponent<Value = ColorPair>>(
&self,
point_a: Point2D<f32, Points>,
point_b: Point2D<f32, Points>,
style: &Style
)

Draws a line between point_a and point_b using style.

Loading content...