Trait gooey_renderer::Renderer

pub trait Renderer: Debug + Send + Sync + Sized + 'static {
    fn size(&self) -> Size2D<f32, Points>;

    fn clip_to(&self, bounds: Rect<f32, Points>) -> Self;

    fn clip_bounds(&self) -> Rect<f32, Points>;

    fn scale(&self) -> Scale<f32, Points, Pixels>;

    fn render_text<F: FallbackComponent<Value = ColorPair>>(
        &self, 
        text: &str, 
        baseline_origin: Point2D<f32, Points>, 
        style: &Style
    );

    fn measure_text(&self, text: &str, style: &Style) -> TextMetrics<Points>;

    fn stroke_rect(&self, rect: &Rect<f32, Points>, style: &Style);

    fn fill_rect<F: FallbackComponent<Value = ColorPair>>(
        &self, 
        rect: &Rect<f32, Points>, 
        style: &Style
    );

    fn stroke_line(
        &self, 
        point_a: Point2D<f32, Points>, 
        point_b: Point2D<f32, Points>, 
        style: &Style
    );

    fn bounds(&self) -> Rect<f32, Points> { ... }
}

Implements drawing APIs.

Required methods

fn size(&self) -> Size2D<f32, Points>

The size of the area being drawn.

fn clip_to(&self, bounds: Rect<f32, Points>) -> Self

Returns a new renderer instance with the state such that each operation executes as if the origin is bounds.origin. The returned instance’s size() should equal bounds.size.

fn clip_bounds(&self) -> Rect<f32, Points>

A [Rect] representing the area being drawn. This rect should be offset relative to the origin of the renderer.

fn scale(&self) -> Scale<f32, Points, Pixels>

The scale when converting between Points and Pixels.

fn render_text<F: FallbackComponent<Value = ColorPair>>(
&self,
text: &str,
baseline_origin: Point2D<f32, Points>,
style: &Style
)

Renders text at baseline_origin with options.

fn measure_text(&self, text: &str, style: &Style) -> TextMetrics<Points>

Measures text using options.

fn stroke_rect(&self, rect: &Rect<f32, Points>, style: &Style)

Strokes the outline of rect using options.

fn fill_rect<F: FallbackComponent<Value = ColorPair>>(
&self,
rect: &Rect<f32, Points>,
style: &Style
)

Fills rect using color.

fn stroke_line(
&self,
point_a: Point2D<f32, Points>,
point_b: Point2D<f32, Points>,
style: &Style
)

Draws a line between point_a and point_b using options.

Loading content...
Provided methods

fn bounds(&self) -> Rect<f32, Points>

A [Rect] representing the area being drawn. Due to how rendering works, the origin is always zero.

Loading content...